Move viewmodel to a new project
