+++ SchoolManagerViewModel/Commands/AddUserCommand.cs
@@ -1,13 +1,13 @@
-﻿using SchoolManagerModel.Entities;
+﻿using System.Diagnostics;
+using System.Windows.Input;
+using SchoolManagerModel.Entities;
 using SchoolManagerModel.Entities.UserModel;
 using SchoolManagerModel.Managers;
 using SchoolManagerModel.Persistence;
 using SchoolManagerModel.Utils;
 using SchoolManagerModel.Validators;
-using System.Diagnostics;
-using System.Windows.Input;
 
-namespace SchoolManagerWPF.ViewModel.Commands;
+namespace SchoolManagerViewModel.Commands;
 
 public class AddUserCommand : ICommand
