+++ SchoolManagerWPF.sln
@@ -1,11 +1,13 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
-VisualStudioVersion = 17.12.35424.110 d17.12
+VisualStudioVersion = 17.12.35424.110
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SchoolManagerWPF", "SchoolManagerWPF\SchoolManagerWPF.csproj", "{656ECE4F-D728-4984-87A8-9C626FF499F0}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SchoolManagerModel", "SchoolManagerModel\SchoolManagerModel\SchoolManagerModel.csproj", "{687F0C90-94B1-424E-8773-4E355BBD9498}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SchoolManagerViewModel", "SchoolManagerViewModel\SchoolManagerViewModel.csproj", "{36BA5627-0FB0-484D-B7E7-C8D96BF8A5E0}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -21,6 +23,10 @@ Global
 		{687F0C90-94B1-424E-8773-4E355BBD9498}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{687F0C90-94B1-424E-8773-4E355BBD9498}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{687F0C90-94B1-424E-8773-4E355BBD9498}.Release|Any CPU.Build.0 = Release|Any CPU
+		{36BA5627-0FB0-484D-B7E7-C8D96BF8A5E0}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{36BA5627-0FB0-484D-B7E7-C8D96BF8A5E0}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{36BA5627-0FB0-484D-B7E7-C8D96BF8A5E0}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{36BA5627-0FB0-484D-B7E7-C8D96BF8A5E0}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
