+++ SchoolManagerWPF/App.xaml
@@ -1,7 +1,7 @@
-﻿<Application x:Class="SchoolManagerWPF.App"
+﻿<Application x:Class="SchoolManagerViewModel.App"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-             xmlns:local="clr-namespace:SchoolManagerWPF"
+             xmlns:local="clr-namespace:SchoolManagerViewModel"
              StartupUri="View/SelectLanguage.xaml">
     <Application.Resources>
         <Style TargetType="{x:Type Window}">
