+++ SchoolManagerWPF/SchoolManagerWPF.csproj
@@ -15,6 +15,7 @@
 
   <ItemGroup>
     <ProjectReference Include="..\SchoolManagerModel\SchoolManagerModel\SchoolManagerModel.csproj" />
+    <ProjectReference Include="..\SchoolManagerViewModel\SchoolManagerViewModel.csproj" />
   </ItemGroup>
 
   <ItemGroup>
