+++ SchoolManagerWPF/View/Controls/CheckBoxList.xaml
@@ -1,4 +1,4 @@
-﻿<UserControl x:Class="SchoolManagerWPF.View.Controls.CheckBoxList"
+﻿<UserControl x:Class="SchoolManagerViewModel.View.Controls.CheckBoxList"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
