+++ SchoolManagerWPF/View/Controls/ValidatedPasswordField.xaml
@@ -1,4 +1,4 @@
-﻿<UserControl x:Class="SchoolManagerWPF.View.Controls.ValidatedPasswordField"
+﻿<UserControl x:Class="SchoolManagerViewModel.View.Controls.ValidatedPasswordField"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
