+++ SchoolManagerWPF/View/Controls/ValidatedPasswordField.xaml.cs
@@ -1,7 +1,7 @@
 ﻿using System.Windows;
 using System.Windows.Controls;
 
-namespace SchoolManagerWPF.View.Controls;
+namespace SchoolManagerViewModel.View.Controls;
 
 public partial class ValidatedPasswordField : UserControl
 {
