+++ SchoolManagerWPF/View/Login.xaml
@@ -1,9 +1,9 @@
-﻿<Window x:Class="SchoolManagerWPF.Login"
+﻿<Window x:Class="SchoolManagerViewModel.Login"
         xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
         xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
         xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-        xmlns:local="clr-namespace:SchoolManagerWPF"
+        xmlns:local="clr-namespace:SchoolManagerViewModel"
         xmlns:resx="clr-namespace:SchoolManagerModel.Resources;assembly=SchoolManagerModel"
         mc:Ignorable="d"
         Title="Login"
