+++ SchoolManagerWPF/View/SelectLanguage.xaml
@@ -1,4 +1,4 @@
-﻿<Window x:Class="SchoolManagerWPF.SelectLanguage"
+﻿<Window x:Class="SchoolManagerViewModel.SelectLanguage"
         xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
         xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
