+++ SchoolManagerWPF/View/StudentUI.xaml
@@ -1,9 +1,9 @@
-﻿<Window x:Class="SchoolManagerWPF.View.StudentUI"
+﻿<Window x:Class="SchoolManagerViewModel.View.StudentUI"
         xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
         xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
         xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-        xmlns:local="clr-namespace:SchoolManagerWPF.View"
+        xmlns:local="clr-namespace:SchoolManagerViewModel.View"
         xmlns:resx="clr-namespace:SchoolManagerModel.Resources;assembly=SchoolManagerModel"
         mc:Ignorable="d"
         Title="StudentUI"
