+++ SchoolManagerWPF/View/TabPages/Roster.xaml
@@ -1,12 +1,12 @@
-﻿<UserControl x:Class="SchoolManagerWPF.View.TabPages.Roster"
+﻿<UserControl x:Class="SchoolManagerViewModel.View.TabPages.Roster"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
              xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-             xmlns:local="clr-namespace:SchoolManagerWPF.View.TabPages"
+             xmlns:local="clr-namespace:SchoolManagerViewModel.View.TabPages"
              xmlns:resx="clr-namespace:SchoolManagerModel.Resources;assembly=SchoolManagerModel"
-             xmlns:controls="clr-namespace:SchoolManagerWPF.View.Controls"
-             xmlns:converters="clr-namespace:SchoolManagerWPF.Converters"
+             xmlns:controls="clr-namespace:SchoolManagerViewModel.View.Controls"
+             xmlns:converters="clr-namespace:SchoolManagerViewModel.Converters"
              mc:Ignorable="d"
              d:DesignHeight="450"
              d:DesignWidth="800">
