+++ SchoolManagerWPF/View/TabPages/Roster.xaml.cs
@@ -1,9 +1,8 @@
-﻿using SchoolManagerWPF.ViewModel;
-using System.ComponentModel;
+﻿using System.ComponentModel;
 using System.Windows;
 using System.Windows.Controls;
 
-namespace SchoolManagerWPF.View.TabPages;
+namespace SchoolManagerViewModel.View.TabPages;
 
 /// <summary>
 /// Interaction logic for Roster.xaml
